syslog: Move syslog stream to libc like other stream implementation
