+++ drivers/syslog/vsyslog.c
@@ -30,8 +30,8 @@
 
 #include <nuttx/arch.h>
 #include <nuttx/init.h>
-#include <nuttx/arch.h>
 #include <nuttx/clock.h>
+#include <nuttx/streams.h>
 
 #include "syslog.h"
 
@@ -84,7 +84,7 @@ int nx_vsyslog(int priority, FAR const IPTR char *fmt, FAR va_list *ap)
    * do the work.
    */
 
-  syslogstream_create(&stream);
+  lib_syslogstream_open(&stream);
 
 #ifdef CONFIG_SYSLOG_TIMESTAMP
   ts.tv_sec = 0;
@@ -228,11 +228,8 @@ int nx_vsyslog(int priority, FAR const IPTR char *fmt, FAR va_list *ap)
   ret += lib_sprintf(&stream.public, "\e[0m");
 #endif
 
-#ifdef CONFIG_SYSLOG_BUFFER
   /* Flush and destroy the syslog stream buffer */
 
-  syslogstream_destroy(&stream);
-#endif
-
+  lib_syslogstream_close(&stream);
   return ret;
 }
