+++ include/nuttx/streams.h
@@ -198,6 +198,21 @@ struct lib_rawsostream_s
   int                    fd;
 };
 
+/* This is a special stream that does buffered character I/O.  NOTE that is
+ * CONFIG_SYSLOG_BUFFER is not defined, it is the same as struct
+ * lib_outstream_s
+ */
+
+struct iob_s;  /* Forward reference */
+
+struct lib_syslogstream_s
+{
+  struct lib_outstream_s public;
+#ifdef CONFIG_SYSLOG_BUFFER
+  FAR struct iob_s *iob;
+#endif
+};
+
 /* LZF compressed stream pipeline */
 
 #ifdef CONFIG_LIBC_LZF
@@ -387,6 +402,45 @@ void lib_lowoutstream(FAR struct lib_outstream_s *lowoutstream);
 void lib_zeroinstream(FAR struct lib_instream_s *zeroinstream);
 void lib_nullinstream(FAR struct lib_instream_s *nullinstream);
 void lib_nulloutstream(FAR struct lib_outstream_s *nulloutstream);
+
+/****************************************************************************
+ * Name: lib_syslogstream_open
+ *
+ * Description:
+ *   Initializes a stream for use with the configured syslog interface.
+ *   Only accessible from with the OS SYSLOG logic.
+ *
+ * Input Parameters:
+ *   stream - User allocated, uninitialized instance of struct
+ *            lib_syslogstream_s to be initialized.
+ *
+ * Returned Value:
+ *   None (User allocated instance initialized).
+ *
+ ****************************************************************************/
+
+void lib_syslogstream_open(FAR struct lib_syslogstream_s *stream);
+
+/****************************************************************************
+ * Name: lib_syslogstream_close
+ *
+ * Description:
+ *   Free resources held by the syslog stream.
+ *
+ * Input Parameters:
+ *   stream - User allocated, uninitialized instance of struct
+ *            lib_syslogstream_s to be initialized.
+ *
+ * Returned Value:
+ *   None (Resources freed).
+ *
+ ****************************************************************************/
+
+#ifdef CONFIG_SYSLOG_BUFFER
+void lib_syslogstream_close(FAR struct lib_syslogstream_s *stream);
+#else
+#  define lib_syslogstream_close(s)
+#endif
 
 /****************************************************************************
  * Name: lib_lzfoutstream
